requirements.txt
streamlit
sentence-transformers
faiss-cpu
groq
python-dotenv
pypdf
numpy